Reduced footprint
Replaced uint64_t with uint32_t where possible.

+++ lib-h3/include/h3_ccu.h
@@ -90,7 +90,7 @@ typedef enum H3_CCU_PLL {
 extern "C" {
 #endif
 
-extern uint64_t h3_ccu_get_pll_rate(ccu_pll_t);
+extern uint32_t h3_ccu_get_pll_rate(ccu_pll_t);
 
 // NDEBUG
 extern void h3_ccu_pll_dump(void);
